requirements.txt
pygame-ce>=2.3.0
pytest>=7.0.0
